+++ sdk/core/azure-core/src/uuid.cpp
@@ -16,36 +16,54 @@ static thread_local std::mt19937_64 randomGenerator(std::random_device{}());
 } // namespace
 #endif
 
+namespace {
+static uint8_t HexCharToByte(char c)
+{
+  if (c >= '0' && c <= '9')
+    return c - '0';
+  if (c >= 'a' && c <= 'f')
+    return c - 'a' + 10;
+  if (c >= 'A' && c <= 'F')
+    return c - 'A' + 10;
+
+  throw std::invalid_argument("Invalid hexadecimal character.");
+}
+
+static std::string ToStringHelper(std::array<uint8_t, 16> uuid)
+{
+  // Guid is 36 characters
+  //  Add one byte for the \0
+  char s[37];
+
+  std::snprintf(
+      s,
+      sizeof(s),
+      "%2.2x%2.2x%2.2x%2.2x-%2.2x%2.2x-%2.2x%2.2x-%2.2x%2.2x-%2.2x%2.2x%2.2x%2.2x%2.2x%2.2x",
+      uuid[0],
+      uuid[1],
+      uuid[2],
+      uuid[3],
+      uuid[4],
+      uuid[5],
+      uuid[6],
+      uuid[7],
+      uuid[8],
+      uuid[9],
+      uuid[10],
+      uuid[11],
+      uuid[12],
+      uuid[13],
+      uuid[14],
+      uuid[15]);
+
+  return std::string(s);
+}
+} // namespace
+
 namespace Azure { namespace Core {
-  std::string Uuid::ToString()
-  {
-    // Guid is 36 characters
-    //  Add one byte for the \0
-    char s[37];
+  std::string Uuid::ToString() const { return ToStringHelper(m_uuid); }
 
-    std::snprintf(
-        s,
-        sizeof(s),
-        "%2.2x%2.2x%2.2x%2.2x-%2.2x%2.2x-%2.2x%2.2x-%2.2x%2.2x-%2.2x%2.2x%2.2x%2.2x%2.2x%2.2x",
-        m_uuid[0],
-        m_uuid[1],
-        m_uuid[2],
-        m_uuid[3],
-        m_uuid[4],
-        m_uuid[5],
-        m_uuid[6],
-        m_uuid[7],
-        m_uuid[8],
-        m_uuid[9],
-        m_uuid[10],
-        m_uuid[11],
-        m_uuid[12],
-        m_uuid[13],
-        m_uuid[14],
-        m_uuid[15]);
-
-    return std::string(s);
-  }
+  std::string Uuid::ToString() { return ToStringHelper(m_uuid); }
 
   Uuid Uuid::CreateUuid()
   {
@@ -88,6 +106,35 @@ namespace Azure { namespace Core {
   {
     Uuid rv{uuid.data()};
     return rv;
+  }
+
+  Uuid Uuid::CreateFromString(std::string const& uuid)
+  {
+    if (uuid.size() != 36)
+    {
+      throw std::invalid_argument("Uuid string must be exactly 36 characters long.");
+    }
+    if (uuid[8] != '-' || uuid[13] != '-' || uuid[18] != '-' || uuid[23] != '-')
+    {
+      throw std::invalid_argument("Only the `8-4-4-4-12` Uuid format is supported.");
+    }
+
+    std::array<uint8_t, UuidSize> uuidArray{};
+
+    for (size_t i = 0, j = 0; j < UuidSize && i < uuid.size();)
+    {
+      if (uuid[i] == '-')
+      {
+        i++;
+        continue; // Skip hyphens
+      }
+
+      uuidArray[j] = (HexCharToByte(uuid[i++]) << 4);
+      uuidArray[j] |= HexCharToByte(uuid[i++]);
+      j++;
+    }
+
+    return Uuid::CreateFromArray(uuidArray);
   }
 
 }} // namespace Azure::Core