+++ sdk/core/azure-core/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features Added
 
 - Request logs to now include the `accept-range`, `content-range`, `range`, `WWW-Authenticate`, `x-ms-date`, `x-ms-error-code`, `x-ms-range`, and `x-ms-version` headers.
+- Added `Uuid::CreateFromString()` method.
+
 ### Breaking Changes
 
 ### Bugs Fixed